moved executive board to members
