+++ src/css/styles2.css
@@ -84,3 +84,105 @@
   color:white;
 }
 
+.become-a-member-b a{
+  text-decoration:none;
+  color:white;
+}
+
+.become-a-member-b:hover{
+  background-color:#0A5471;
+  opacity:0.8;
+}
+
+.become-a-member-b:focus{
+  outline:none;
+  box-shadow:none;
+}
+
+.become-a-member-y a{
+  text-decoration:none;
+  color:white;
+}
+
+.become-a-member-y:hover{
+  background-color:rgba(227, 188, 89, 1);
+  opacity:0.8;
+}
+
+.become-a-member-y:focus{
+  outline:none;
+  box-shadow:none;
+}
+
+#membership-content {
+  margin-top:5em;
+  display:flex;
+  justify-content:space-evenly;
+  margin-bottom:10em;
+}
+
+#membership-card {
+  width:48%;
+  padding:2em;
+  background-color:#0A5471;
+}
+
+#membership-card h1{
+  color:white;
+  text-align:center;
+  margin-bottom:1em;
+}
+
+#membership-card a{
+  color:white;
+  text-align:center;
+  font-size: calc(1.275rem + 0.3vw);
+  text-decoration:none;
+  margin-bottom:.5em;
+}
+
+#membership-card a:hover{
+  text-decoration:underline;
+}
+
+.sub-btn {
+  background-color:#0A5471;
+  border:#0A5471;
+  display:block;
+  margin:auto;
+  width:30%;
+  margin-top:2em;
+  outline:none;
+  box-shadow:none;
+}
+
+.sub-btn:hover{
+  background-color:#0A5471;
+  opacity:.8;
+  outline:none;
+  box-shadow:none;
+}
+
+.sub-btn:active{
+  background-color:#0A5471;
+  outline:none;
+  box-shadow:none;
+}
+
+.sub-btn:focus{
+  background-color:#0A5471;
+  outline:none;
+  box-shadow:none;
+}
+
+#sub-btn-sec{
+  display:block;
+  margin:auto;
+  width:30%;
+}
+
+.mailing-form{
+  width:60%;
+  display:block;
+  margin:auto;
+}
